Flatten client hierarchy and remove Raw prefix
Remove the unified Client/WebsocketClient classes that combined sync and
async via use_async flag. Promote the Raw* implementations to be the
public API: Client, AsyncClient, WebsocketClient, AsyncWebsocketClient.
Drop the async_ prefix from async client methods since they now live on
separate classes. Move URL scheme validation into base classes.
