+++ homeassistant_api/__init__.py
@@ -2,6 +2,7 @@
 
 __all__ = (
     "Client",
+    "AsyncClient",
     "State",
     "Context",
     "Domain",
@@ -12,6 +13,7 @@ __all__ = (
     "Event",
     "LogbookEntry",
     "WebsocketClient",
+    "AsyncWebsocketClient",
     "AuthInvalid",
     "AuthOk",
     "AuthRequired",
@@ -21,6 +23,8 @@ __all__ = (
     "EventResponse",
 )
 
+from .asyncclient import AsyncClient
+from .asyncwebsocket import AsyncWebsocketClient
 from .client import Client
 from .models.domains import Domain, Service
 from .models.entity import Entity, Group
