+++ homeassistant_api/asyncclient.py
@@ -8,7 +8,6 @@ import logging
 from datetime import datetime
 from posixpath import join
 from typing import (
-    TYPE_CHECKING,
     Any,
     AsyncGenerator,
     Dict,
@@ -26,20 +25,15 @@ import aiohttp_client_cache.session
 from .errors import BadTemplateError, RequestError, RequestTimeoutError
 from .models import Domain, Entity, Event, Group, History, LogbookEntry, State
 from .processing import AsyncResponseType, Processing
-from .rawbaseclient import RawBaseClient
+from .baseclient import BaseClient
 from .utils import JSONType, prepare_entity_id
-
-if TYPE_CHECKING:
-    from homeassistant_api import Client
-else:
-    Client = None  # pylint: disable=invalid-name
 
 logger = logging.getLogger(__name__)
 
 
-class RawAsyncClient(RawBaseClient):
+class AsyncClient(BaseClient):
     """
-    The async equivalent of :py:class:`RawClient`
+    The async equivalent of :py:class:`Client`
 
     :param api_url: The location of the api endpoint. e.g. :code:`http://localhost:8123/api` Required.
     :param token: The refresh or long lived access token to authenticate your requests. Required.
@@ -61,7 +55,7 @@ class RawAsyncClient(RawBaseClient):
         verify_ssl: bool = True,
         **kwargs,
     ):
-        RawBaseClient.__init__(self, *args, **kwargs)
+        BaseClient.__init__(self, *args, **kwargs)
         connector = aiohttp.TCPConnector(verify_ssl=False) if not verify_ssl else None
         if async_cache_session is False:
             self.async_cache_session = aiohttp.ClientSession(connector=connector)
@@ -81,7 +75,7 @@ class RawAsyncClient(RawBaseClient):
             "Entering cached async requests session %r", self.async_cache_session
         )
         await self.async_cache_session.__aenter__()
-        await self.async_check_api_running()
+        await self.check_api_running()
         return self
 
     async def __aexit__(self, _, __, ___):
@@ -89,7 +83,7 @@ class RawAsyncClient(RawBaseClient):
         await self.async_cache_session.close()
 
     # Very important request function
-    async def async_request(
+    async def request(
         self,
         path: str,
         *,
@@ -102,7 +96,7 @@ class RawAsyncClient(RawBaseClient):
         try:
             if self.global_request_kwargs is not None:
                 kwargs.update(self.global_request_kwargs)
-            return await self.async_response_logic(
+            return await self.response_logic(
                 await self.async_cache_session.request(
                     method,
                     self.endpoint(path) + f"?{params}" * bool(params),
@@ -116,27 +110,45 @@ class RawAsyncClient(RawBaseClient):
                 self.endpoint(path) + f"?{params}" * bool(params),
             ) from err
 
+    async def _dict_request(self, *args: Any, **kwargs: Any) -> dict:
+        data = await self.request(*args, **kwargs)
+        if not isinstance(data, dict):
+            raise TypeError
+        return data
+
+    async def _list_request(self, *args: Any, **kwargs: Any) -> list:
+        data = await self.request(*args, **kwargs)
+        if not isinstance(data, list):
+            raise TypeError
+        return data
+
+    async def _str_request(self, *args: Any, **kwargs: Any) -> str:
+        data = await self.request(*args, **kwargs)
+        if not isinstance(data, str):
+            raise TypeError
+        return data
+
     @staticmethod
-    async def async_response_logic(response: AsyncResponseType) -> Any:
+    async def response_logic(response: AsyncResponseType) -> Any:
         """Processes custom mimetype content asyncronously."""
         return await Processing(response=response).process()
 
     # API information methods
-    async def async_get_error_log(self) -> str:
+    async def get_error_log(self) -> str:
         """
         Returns the server error log as a string.
         :code:`GET /api/error_log`
         """
-        return cast(str, await self.async_request("error_log"))
+        return cast(str, await self.request("error_log"))
 
-    async def async_get_config(self) -> dict[str, JSONType]:
+    async def get_config(self) -> dict[str, JSONType]:
         """
         Returns the yaml configuration of homeassistant.
         :code:`GET /api/config`
         """
-        return cast(dict[str, JSONType], await self.async_request("config"))
+        return cast(dict[str, JSONType], await self.request("config"))
 
-    async def async_get_logbook_entries(
+    async def get_logbook_entries(
         self,
         *args,
         **kwargs,
@@ -146,13 +158,13 @@ class RawAsyncClient(RawBaseClient):
         :code:`GET /api/logbook/<timestamp>`
         """
         params, url = self.prepare_get_logbook_entry_params(*args, **kwargs)
-        data = await self.async_request(
+        data = await self.request(
             url, params=self.construct_params(cast(Dict[str, Optional[str]], params))
         )
         for entry in data:
             yield LogbookEntry.model_validate(entry)
 
-    async def async_get_entity_histories(
+    async def get_entity_histories(
         self,
         entities: Optional[Tuple[Entity, ...]] = None,
         start_timestamp: Optional[datetime] = None,
@@ -170,14 +182,14 @@ class RawAsyncClient(RawBaseClient):
             end_timestamp=end_timestamp,
             significant_changes_only=significant_changes_only,
         )
-        data = await self.async_request(
+        data = await self.request(
             url,
             params=self.construct_params(params),
         )
         for states in data:
             yield History.model_validate({"states": states})
 
-    async def async_get_rendered_template(self, template: str) -> str:
+    async def get_rendered_template(self, template: str) -> str:
         """
         Renders a given Jinja2 template string with Home Assistant context data.
         :code:`POST /api/template`
@@ -185,7 +197,7 @@ class RawAsyncClient(RawBaseClient):
         try:
             return cast(
                 str,
-                await self.async_request(
+                await self.request(
                     "template",
                     json=dict(template=template),
                     method="POST",
@@ -198,12 +210,12 @@ class RawAsyncClient(RawBaseClient):
             ) from err
 
     # API check methods
-    async def async_check_api_config(self) -> bool:
+    async def check_api_config(self) -> bool:
         """
         Asks Home Assistant to validate its configuration file and returns true/false.
         :code:`POST /api/config/core/check_config`
         """
-        res = await self.async_request("config/core/check_config", method="POST")
+        res = await self.request("config/core/check_config", method="POST")
         res = cast(Dict[Any, Any], res)
         valid = {"valid": True, "invalid": False}.get(
             cast(
@@ -214,29 +226,29 @@ class RawAsyncClient(RawBaseClient):
         )
         return valid
 
-    async def async_check_api_running(self) -> bool:
+    async def check_api_running(self) -> bool:
         """
         Asks Home Assistant if its running.
         :code:`GET /api/`
         """
-        res = cast(Dict[Any, Any], await self.async_request(""))
+        res = cast(Dict[Any, Any], await self.request(""))
         return res.get("message") == "API running."
 
     # Entity methods
-    async def async_get_entities(self) -> Dict[str, Group]:
+    async def get_entities(self) -> Dict[str, Group]:
         """
         Fetches all entities from the api.
         :code:`GET /api/states`
         """
         entities: Dict[str, Group] = {}
-        for state in await self.async_get_states():
+        for state in await self.get_states():
             group_id, entity_slug = state.entity_id.split(".")
             if group_id not in entities:
                 entities[group_id] = Group(group_id=group_id, _client=self)  # type: ignore[arg-type]
             entities[group_id]._add_entity(entity_slug, state)
         return entities
 
-    async def async_get_entity(
+    async def get_entity(
         self,
         group_id: Optional[str] = None,
         slug: Optional[str] = None,
@@ -247,9 +259,9 @@ class RawAsyncClient(RawBaseClient):
         :code:`GET /api/states/<entity_id>`
         """
         if group_id is not None and slug is not None:
-            state = await self.async_get_state(group_id=group_id, slug=slug)
+            state = await self.get_state(group_id=group_id, slug=slug)
         elif entity_id is not None:
-            state = await self.async_get_state(entity_id=entity_id)
+            state = await self.get_state(entity_id=entity_id)
         else:
             help_msg = (
                 "Use keyword arguments to pass entity_id. "
@@ -264,27 +276,29 @@ class RawAsyncClient(RawBaseClient):
         return group.get_entity(entity_slug)
 
     # Services and domain methods
-    async def async_get_domains(self) -> Dict[str, Domain]:
+    async def get_domains(self) -> Dict[str, Domain]:
         """
         Fetches all :py:class:`Service` 's from the API.
         :code:`GET /api/services`
         """
-        data = await self.async_request("services")
+        data = await self.request("services")
         domains = map(
-            lambda json: Domain.from_json_with_client(json, client=cast(Client, self)),
+            lambda json: Domain.from_json_with_client(
+                json, client=cast("AsyncClient", self)
+            ),
             cast(Tuple[dict[str, JSONType], ...], data),
         )
         return {domain.domain_id: domain for domain in domains}
 
-    async def async_get_domain(self, domain_id: str) -> Optional[Domain]:
+    async def get_domain(self, domain_id: str) -> Optional[Domain]:
         """
         Fetches all :py:class:`Service`'s under a particular service :py:class:`Domain`.
         Uses cached data from :py:meth:`get_domains` if available.
         """
-        domains = await self.async_get_domains()
+        domains = await self.get_domains()
         return domains.get(domain_id)
 
-    async def async_trigger_service(
+    async def trigger_service(
         self,
         domain: str,
         service: str,
@@ -294,14 +308,14 @@ class RawAsyncClient(RawBaseClient):
         Tells Home Assistant to trigger a service, returns all states changed while in the process of being called.
         :code:`POST /api/services/<domain>/<service>`
         """
-        data = await self.async_request(
+        data = await self.request(
             f"services/{domain}/{service}",
             method="POST",
             json=service_data,
         )
         return tuple(map(State.from_json, cast(List[Dict[Any, Any]], data)))
 
-    async def async_trigger_service_with_response(
+    async def trigger_service_with_response(
         self,
         domain: str,
         service: str,
@@ -315,7 +329,7 @@ class RawAsyncClient(RawBaseClient):
         """
         data = cast(
             dict[str, dict[str, JSONType]],
-            await self.async_request(
+            await self.request(
                 join("services", domain, service) + "?return_response",
                 method="POST",
                 json=service_data,
@@ -330,7 +344,7 @@ class RawAsyncClient(RawBaseClient):
         return states, data.get("service_response", {})
 
     # EntityState methods
-    async def async_get_state(  # pylint: disable=duplicate-code
+    async def get_state(  # pylint: disable=duplicate-code
         self,
         *,
         entity_id: Optional[str] = None,
@@ -346,10 +360,10 @@ class RawAsyncClient(RawBaseClient):
             slug=slug,
             entity_id=entity_id,
         )
-        data = await self.async_request(join("states", target_entity_id))
+        data = await self.request(join("states", target_entity_id))
         return State.from_json(cast(Dict[Any, Any], data))
 
-    async def async_set_state(  # pylint: disable=duplicate-code
+    async def set_state(  # pylint: disable=duplicate-code
         self,
         state: State,
     ) -> State:
@@ -358,63 +372,63 @@ class RawAsyncClient(RawBaseClient):
         To communicate with the device, use :py:meth:`Service.trigger` or :py:meth:`Service.async_trigger`.
         :code:`POST /api/states/<entity_id>`
         """
-        data = await self.async_request(
+        data = await self.request(
             join("states", state.entity_id),
             method="POST",
             json=json.loads(state.model_dump_json()),
         )
         return State.from_json(cast(Dict[Any, Any], data))
 
-    async def async_get_states(self) -> Tuple[State, ...]:
+    async def get_states(self) -> Tuple[State, ...]:
         """
         Gets the states of all entities within homeassistant.
         :code:`GET /api/states`
         """
-        data = await self.async_request("states")
+        data = await self.request("states")
         return tuple(map(State.from_json, cast(List[Dict[Any, Any]], data)))
 
     # Event methods
-    async def async_get_events(self) -> Tuple[Event, ...]:
+    async def get_events(self) -> Tuple[Event, ...]:
         """
         Gets the Events that happen within homeassistant
         :code:`GET /api/events`
         """
-        data = await self.async_request("events")
+        data = await self.request("events")
         return tuple(
             map(
                 lambda json: Event.from_json_with_client(
-                    json, client=cast(Client, self)
+                    json, client=cast("AsyncClient", self)
                 ),
                 cast(List[dict[str, JSONType]], data),
             )
         )
 
-    async def async_get_event(self, name: str) -> Optional[Event]:
+    async def get_event(self, name: str) -> Optional[Event]:
         """
         Gets the :py:class:`Event` with the specified name if it has at least one listener.
         Uses cached data from :py:meth:`get_events` if available.
         """
-        for event in await self.async_get_events():
+        for event in await self.get_events():
             if event.event == name.strip().lower():
                 return event
         return None
 
-    async def async_fire_event(self, event_type: str, **event_data: Any) -> str:
+    async def fire_event(self, event_type: str, **event_data: Any) -> str:
         """
         Fires a given event_type within homeassistant. Must be an existing event_type.
         :code:`POST /api/events/<event_type>`
         """
-        data = await self.async_request(
+        data = await self.request(
             join("events", event_type),
             method="POST",
             json=event_data,
         )
         return cast(str, data.get("message", "No message provided"))
 
-    async def async_get_components(self) -> Tuple[str, ...]:
+    async def get_components(self) -> Tuple[str, ...]:
         """
         Returns a tuple of all registered components.
         :code:`GET /api/components`
         """
-        data = await self.async_request("components")
+        data = await self.request("components")
         return tuple(cast(List[str], data))
