+++ tests/test_endpoints.py
@@ -5,12 +5,11 @@ from datetime import datetime
 
 import pytest
 
-from homeassistant_api import Client
+from homeassistant_api import AsyncClient, AsyncWebsocketClient, Client, WebsocketClient
 from homeassistant_api.errors import RequestError
 from homeassistant_api.models import ConfigEntryDisabler
 from homeassistant_api.models.events import Event
 from homeassistant_api.models.states import State
-from homeassistant_api.websocket import WebsocketClient
 
 
 def test_get_error_log(cached_client: Client) -> None:
@@ -18,9 +17,9 @@ def test_get_error_log(cached_client: Client) -> None:
     assert cached_client.get_error_log()
 
 
-async def test_async_get_error_log(async_cached_client: Client) -> None:
+async def test_async_get_error_log(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/error_log` endpoint."""
-    assert await async_cached_client.async_get_error_log()
+    assert await async_cached_client.get_error_log()
 
 
 def test_get_config(cached_client: Client) -> None:
@@ -28,9 +27,9 @@ def test_get_config(cached_client: Client) -> None:
     assert cached_client.get_config().get("state") in {"RUNNING", "NOT_RUNNING"}
 
 
-async def test_async_get_config(async_cached_client: Client) -> None:
+async def test_async_get_config(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/config` endpoint."""
-    assert (await async_cached_client.async_get_config()).get("state") in {
+    assert (await async_cached_client.get_config()).get("state") in {
         "RUNNING",
         "NOT_RUNNING",
     }
@@ -46,9 +45,9 @@ def test_get_logbook_entries(cached_client: Client) -> None:
         assert entry
 
 
-async def test_async_get_logbook_entries(async_cached_client: Client) -> None:
+async def test_async_get_logbook_entries(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/logbook/<timestamp>` endpoint."""
-    async for entry in async_cached_client.async_get_logbook_entries(
+    async for entry in async_cached_client.get_logbook_entries(
         filter_entities="sun.red_sun",
         start_timestamp=datetime(2020, 1, 1),
         end_timestamp=datetime.now(),
@@ -61,9 +60,9 @@ def test_get_entity(cached_client: Client) -> None:
     assert cached_client.get_entity(entity_id="sun.sun")
 
 
-async def test_async_get_entity(async_cached_client: Client) -> None:
+async def test_async_get_entity(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/states/<entity_id>` endpoint."""
-    assert await async_cached_client.async_get_entity(entity_id="sun.sun")
+    assert await async_cached_client.get_entity(entity_id="sun.sun")
 
 
 def test_get_entity_histories(cached_client: Client) -> None:
@@ -83,13 +82,12 @@ def test_get_entity_histories(cached_client: Client) -> None:
     assert isinstance(histories[0].states[0], State)
 
 
-async def test_async_get_entity_histories(async_cached_client: Client) -> None:
+async def test_async_get_entity_histories(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/history/period/<timestamp>` endpoint."""
-    sun = await async_cached_client.async_get_entity(entity_id="sun.sun")
+    sun = await async_cached_client.get_entity(entity_id="sun.sun")
     assert sun is not None
     histories = [
-        history
-        async for history in async_cached_client.async_get_entity_histories((sun,))
+        history async for history in async_cached_client.get_entity_histories((sun,))
     ]
     assert histories, "No history found."
     assert histories[0].states, "No states in entity history found."
@@ -107,9 +105,9 @@ def test_get_rendered_template(cached_client: Client) -> None:
     }
 
 
-async def test_async_get_rendered_template(async_cached_client: Client) -> None:
+async def test_async_get_rendered_template(async_cached_client: AsyncClient) -> None:
     """Tests the `POST /api/template` endpoint."""
-    rendered_template = await async_cached_client.async_get_rendered_template(
+    rendered_template = await async_cached_client.get_rendered_template(
         'The sun is {{ states("sun.sun").replace("_", " the ") }}.'
     )
     assert rendered_template in {
@@ -130,10 +128,10 @@ def test_websocket_get_rendered_template(websocket_client: WebsocketClient) -> N
 
 
 async def test_async_websocket_get_rendered_template(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "render_template"` websocket command."""
-    rendered_template = await async_websocket_client.async_get_rendered_template(
+    rendered_template = await async_websocket_client.get_rendered_template(
         'The sun is {{ states("sun.sun").replace("_", " the ") }}.'
     )
     assert rendered_template in {
@@ -147,9 +145,9 @@ def test_check_api_config(cached_client: Client) -> None:
     assert cached_client.check_api_config()
 
 
-async def test_async_check_api_config(async_cached_client: Client) -> None:
+async def test_async_check_api_config(async_cached_client: AsyncClient) -> None:
     """Tests the `POST /api/config/core/check_config` endpoint."""
-    assert await async_cached_client.async_check_api_config()
+    assert await async_cached_client.check_api_config()
 
 
 def test_get_entities(cached_client: Client) -> None:
@@ -158,9 +156,9 @@ def test_get_entities(cached_client: Client) -> None:
     assert "sun" in entities
 
 
-async def test_async_get_entities(async_cached_client: Client) -> None:
+async def test_async_get_entities(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/states` endpoint."""
-    entities = await async_cached_client.async_get_entities()
+    entities = await async_cached_client.get_entities()
     assert "sun" in entities
 
 
@@ -172,10 +170,10 @@ def test_websocket_get_config(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_websocket_get_config(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "get_config"` websocket command."""
-    config = await async_websocket_client.async_get_config()
+    config = await async_websocket_client.get_config()
     assert isinstance(config, dict)
     assert config.get("state") in {"RUNNING", "NOT_RUNNING"}
 
@@ -210,48 +208,48 @@ def test_websocket_get_entity_no_args(websocket_client: WebsocketClient) -> None
 
 
 async def test_async_websocket_get_state(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
-    """Tests async WebsocketClient.async_get_state with entity_id."""
-    state = await async_websocket_client.async_get_state(entity_id="sun.sun")
+    """Tests AsyncWebsocketClient.get_state with entity_id."""
+    state = await async_websocket_client.get_state(entity_id="sun.sun")
     assert state.entity_id == "sun.sun"
     assert state.state in {"above_horizon", "below_horizon"}
 
 
 async def test_async_websocket_get_entity_by_group_slug(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
-    """Tests async WebsocketClient.async_get_entity with group_id and slug."""
-    entity = await async_websocket_client.async_get_entity(group_id="sun", slug="sun")
+    """Tests AsyncWebsocketClient.get_entity with group_id and slug."""
+    entity = await async_websocket_client.get_entity(group_id="sun", slug="sun")
     assert entity is not None
     assert entity.entity_id == "sun.sun"
 
 
 async def test_async_websocket_get_entity_by_entity_id(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
-    """Tests async WebsocketClient.async_get_entity with entity_id."""
-    entity = await async_websocket_client.async_get_entity(entity_id="sun.sun")
+    """Tests AsyncWebsocketClient.get_entity with entity_id."""
+    entity = await async_websocket_client.get_entity(entity_id="sun.sun")
     assert entity is not None
     assert entity.entity_id == "sun.sun"
 
 
 async def test_async_websocket_get_entity_no_args(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
-    """Tests async WebsocketClient.async_get_entity raises ValueError with no arguments."""
+    """Tests AsyncWebsocketClient.get_entity raises ValueError with no arguments."""
     with pytest.raises(
         ValueError, match="Neither group_id and slug or entity_id provided"
     ):
-        await async_websocket_client.async_get_entity()
+        await async_websocket_client.get_entity()
 
 
 async def test_async_websocket_get_state_not_found(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
-    """Tests async WebsocketClient.async_get_state raises ValueError for nonexistent entity."""
+    """Tests AsyncWebsocketClient.get_state raises ValueError for nonexistent entity."""
     with pytest.raises(ValueError, match="not found"):
-        await async_websocket_client.async_get_state(
+        await async_websocket_client.get_state(
             entity_id="fake.nonexistent_entity_12345"
         )
 
@@ -269,10 +267,10 @@ def test_websocket_get_entities(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_websocket_get_entities(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "get_entities"` websocket command."""
-    entities = await async_websocket_client.async_get_entities()
+    entities = await async_websocket_client.get_entities()
     assert "sun" in entities
 
 
@@ -282,9 +280,9 @@ def test_get_domains(cached_client: Client) -> None:
     assert "homeassistant" in domains
 
 
-async def test_async_get_domains(async_cached_client: Client) -> None:
+async def test_async_get_domains(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/services` endpoint."""
-    domains = await async_cached_client.async_get_domains()
+    domains = await async_cached_client.get_domains()
     assert "homeassistant" in domains
 
 
@@ -295,10 +293,10 @@ def test_websocket_get_domains(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_websocket_get_domains(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "get_domains"` websocket command."""
-    domains = await async_websocket_client.async_get_domains()
+    domains = await async_websocket_client.get_domains()
     assert "homeassistant" in domains
 
 
@@ -309,9 +307,9 @@ def test_get_domain(cached_client: Client) -> None:
     assert domain.services
 
 
-async def test_async_get_domain(async_cached_client: Client) -> None:
+async def test_async_get_domain(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/services` endpoint."""
-    domain = await async_cached_client.async_get_domain("homeassistant")
+    domain = await async_cached_client.get_domain("homeassistant")
     assert domain is not None
     assert domain.services
 
@@ -324,10 +322,10 @@ def test_websocket_get_domain(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_websocket_get_domain(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "get_domain"` websocket command."""
-    domain = await async_websocket_client.async_get_domain("homeassistant")
+    domain = await async_websocket_client.get_domain("homeassistant")
     assert domain is not None
     assert domain.services
 
@@ -340,10 +338,10 @@ def test_get_nonuser_flows_in_progress(websocket_client: WebsocketClient) -> Non
 
 
 async def test_async_get_nonuser_flows_in_progress(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "config_entries/flow/progress"` websocket command."""
-    flows = await async_websocket_client.async_get_nonuser_flows_in_progress()
+    flows = await async_websocket_client.get_nonuser_flows_in_progress()
     assert not flows
 
 
@@ -369,20 +367,20 @@ def test_disable_enable_config_entry(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_disable_enable_config_entry(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "config_entries/disable"` websocket command."""
-    entry = (await async_websocket_client.async_get_config_entries())[0]
+    entry = (await async_websocket_client.get_config_entries())[0]
     assert entry.disabled_by is None
 
-    await async_websocket_client.async_disable_config_entry(entry.entry_id)
+    await async_websocket_client.disable_config_entry(entry.entry_id)
 
-    disabled_entry = (await async_websocket_client.async_get_config_entries())[0]
+    disabled_entry = (await async_websocket_client.get_config_entries())[0]
     assert disabled_entry.disabled_by is ConfigEntryDisabler.USER
 
-    await async_websocket_client.async_enable_config_entry(entry.entry_id)
+    await async_websocket_client.enable_config_entry(entry.entry_id)
 
-    enabled_entry = (await async_websocket_client.async_get_config_entries())[0]
+    enabled_entry = (await async_websocket_client.get_config_entries())[0]
     assert enabled_entry.disabled_by is None
 
 
@@ -394,11 +392,11 @@ def test_ignore_config_flow(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_ignore_config_flow(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "config_entries/ignore_flow"` websocket command."""
     with pytest.raises(RequestError, match="Config entry not found"):
-        await async_websocket_client.async_ignore_config_flow("", "")
+        await async_websocket_client.ignore_config_flow("", "")
 
 
 def test_get_config_entries(websocket_client: WebsocketClient) -> None:
@@ -431,10 +429,10 @@ def test_get_config_entries(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_get_config_entries(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "config_entries/get"` websocket command."""
-    entries = await async_websocket_client.async_get_config_entries()
+    entries = await async_websocket_client.get_config_entries()
     assert len(entries) == 4
 
     sun = entries[0]
@@ -456,12 +454,12 @@ def test_get_entry_subentries(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_get_entry_subentries(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "config_entries/subentries/list"` websocket command."""
-    sun = (await async_websocket_client.async_get_config_entries())[0]
+    sun = (await async_websocket_client.get_config_entries())[0]
     assert sun
-    assert not await async_websocket_client.async_get_entry_subentries(sun.entry_id)
+    assert not await async_websocket_client.get_entry_subentries(sun.entry_id)
 
 
 def test_delete_entry_subentry(websocket_client: WebsocketClient) -> None:
@@ -472,11 +470,11 @@ def test_delete_entry_subentry(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_delete_entry_subentry(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "config_entries/subentries/delete"` websocket command."""
     with pytest.raises(RequestError, match="Config entry not found"):
-        await async_websocket_client.async_delete_entry_subentry("", "")
+        await async_websocket_client.delete_entry_subentry("", "")
 
 
 def test_trigger_service(cached_client: Client) -> None:
@@ -491,9 +489,9 @@ def test_trigger_service(cached_client: Client) -> None:
     assert isinstance(resp, tuple)
 
 
-async def test_async_trigger_service(async_cached_client: Client) -> None:
+async def test_async_trigger_service(async_cached_client: AsyncClient) -> None:
     """Tests the `POST /api/services/<domain>/<service>` endpoint."""
-    notify = await async_cached_client.async_get_domain("notify")
+    notify = await async_cached_client.get_domain("notify")
     assert notify is not None
     resp = await notify.persistent_notification(
         message="Your API Test Suite just said hello!",
@@ -514,10 +512,10 @@ def test_websocket_trigger_service(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_websocket_trigger_service(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "trigger_service"` websocket command."""
-    notify = await async_websocket_client.async_get_domain("notify")
+    notify = await async_websocket_client.get_domain("notify")
     assert notify is not None
     resp = await notify.persistent_notification(
         message="Your API Test Suite just said hello!", title="Test Suite Notifcation"
@@ -553,9 +551,11 @@ def test_trigger_service_with_response(cached_client: Client) -> None:
     assert data is not None
 
 
-async def test_async_trigger_service_with_response(async_cached_client: Client) -> None:
+async def test_async_trigger_service_with_response(
+    async_cached_client: AsyncClient,
+) -> None:
     """Tests the `POST /api/services/<domain>/<service>?return_response` endpoint."""
-    weather = await async_cached_client.async_get_domain("weather")
+    weather = await async_cached_client.get_domain("weather")
     assert weather is not None
     changed_states, data = await weather.get_forecasts(
         entity_id="weather.forecast_home",
@@ -579,10 +579,10 @@ def test_websocket_trigger_service_with_response(
 
 
 async def test_async_websocket_trigger_service_with_response(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "trigger_service_with_response"` websocket command."""
-    weather = await async_websocket_client.async_get_domain("weather")
+    weather = await async_websocket_client.get_domain("weather")
     assert weather is not None
     data = await weather.get_forecasts(
         entity_id="weather.forecast_home",
@@ -599,9 +599,9 @@ def test_get_states(cached_client: Client) -> None:
         assert isinstance(state, State)
 
 
-async def test_async_get_states(async_cached_client: Client) -> None:
+async def test_async_get_states(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/states` endpoint."""
-    states = await async_cached_client.async_get_states()
+    states = await async_cached_client.get_states()
     for state in states:
         assert isinstance(state, State)
 
@@ -614,10 +614,10 @@ def test_websocket_get_states(websocket_client: WebsocketClient) -> None:
 
 
 async def test_async_websocket_get_states(
-    async_websocket_client: WebsocketClient,
+    async_websocket_client: AsyncWebsocketClient,
 ) -> None:
     """Tests the `"type": "get_states"` websocket command."""
-    states = await async_websocket_client.async_get_states()
+    states = await async_websocket_client.get_states()
     for state in states:
         assert isinstance(state, State)
 
@@ -628,9 +628,9 @@ def test_get_state(cached_client: Client) -> None:
     assert state.state in {"above_horizon", "below_horizon"}
 
 
-async def test_async_get_state(async_cached_client: Client) -> None:
+async def test_async_get_state(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/states/<entity_id>` endpoint."""
-    state = await async_cached_client.async_get_state(entity_id="sun.sun")
+    state = await async_cached_client.get_state(entity_id="sun.sun")
     assert state.state in {"above_horizon", "below_horizon"}
 
 
@@ -642,9 +642,9 @@ def test_set_state(cached_client: Client) -> None:
     assert state.state == "beyond_our_solar_system"
 
 
-async def test_async_set_state(async_cached_client: Client) -> None:
+async def test_async_set_state(async_cached_client: AsyncClient) -> None:
     """Tests the `POST /api/states/<entity_id>` endpoint."""
-    state = await async_cached_client.async_set_state(
+    state = await async_cached_client.set_state(
         State(state="beyond_our_solar_system", entity_id="sun.red_sun")
     )
     assert state.state == "beyond_our_solar_system"
@@ -657,9 +657,9 @@ def test_get_events(cached_client: Client) -> None:
         assert isinstance(event, Event)
 
 
-async def test_async_get_events(async_cached_client: Client) -> None:
+async def test_async_get_events(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/events` endpoint."""
-    events = await async_cached_client.async_get_events()
+    events = await async_cached_client.get_events()
     for event in events:
         assert isinstance(event, Event)
 
@@ -670,9 +670,9 @@ def test_fire_event(cached_client: Client) -> None:
     assert data == "Event my_new_event fired."
 
 
-async def test_async_fire_event(async_cached_client: Client) -> None:
+async def test_async_fire_event(async_cached_client: AsyncClient) -> None:
     """Tests the `POST /api/events/<event_type>` endpoint."""
-    data = await async_cached_client.async_fire_event("my_new_event", parameter="123")
+    data = await async_cached_client.fire_event("my_new_event", parameter="123")
     assert data == "Event my_new_event fired."
 
 
@@ -682,7 +682,7 @@ def test_get_components(cached_client: Client) -> None:
     assert "person" in components
 
 
-async def test_async_get_components(async_cached_client: Client) -> None:
+async def test_async_get_components(async_cached_client: AsyncClient) -> None:
     """Tests the `GET /api/components` endpoint."""
-    components = await async_cached_client.async_get_components()
+    components = await async_cached_client.get_components()
     assert "person" in components
