+++ tests/test_models.py
@@ -5,7 +5,7 @@ from datetime import datetime
 
 import pytest
 
-from homeassistant_api import Client, Domain
+from homeassistant_api import AsyncClient, Client, Domain
 from homeassistant_api.models.events import Event
 from homeassistant_api.models.states import State
 
@@ -21,8 +21,8 @@ def test_entity_get_entity(cached_client: Client) -> None:
         assert person.thispersondoesnotexistplease
 
 
-async def test_async_entity_get_entity(async_cached_client: Client) -> None:
-    person_test_suite = await async_cached_client.async_get_entity(
+async def test_async_entity_get_entity(async_cached_client: AsyncClient) -> None:
+    person_test_suite = await async_cached_client.get_entity(
         group_id="person",
         slug="test_user",
     )
@@ -44,8 +44,8 @@ def test_entity_update_state(cached_client: Client) -> None:
     assert new_state.state == "In the palm of your hand."
 
 
-async def test_async_entity_update_state(async_cached_client: Client) -> None:
-    entity = await async_cached_client.async_get_entity(group_id="sun", slug="red_sun")
+async def test_async_entity_update_state(async_cached_client: AsyncClient) -> None:
+    entity = await async_cached_client.get_entity(group_id="sun", slug="red_sun")
     assert entity is not None
     entity.state.state = "In the palm of my hand."
     new_state = await entity.async_update_state()
@@ -58,10 +58,8 @@ def test_get_event(cached_client: Client) -> None:
     assert event is None
 
 
-async def test_async_get_event(async_cached_client: Client) -> None:
-    event = await async_cached_client.async_get_event(
-        "my_favorite_candy_is_mike_and_ikes"
-    )
+async def test_async_get_event(async_cached_client: AsyncClient) -> None:
+    event = await async_cached_client.get_event("my_favorite_candy_is_mike_and_ikes")
     assert event is None
 
 
@@ -71,8 +69,8 @@ def test_fire_event(cached_client: Client) -> None:
     assert event.fire() == "Event core_config_updated fired."
 
 
-async def test_async_fire_event(async_cached_client: Client) -> None:
-    event = await async_cached_client.async_get_event("core_config_updated")
+async def test_async_fire_event(async_cached_client: AsyncClient) -> None:
+    event = await async_cached_client.get_event("core_config_updated")
     assert event is not None
     assert await event.async_fire() == "Event core_config_updated fired."
 
@@ -85,8 +83,8 @@ def test_get_domain(cached_client: Client) -> None:
         assert notify.thisservicedoesnotexistplease
 
 
-async def test_async_get_domains(async_cached_client: Client) -> None:
-    notify = await async_cached_client.async_get_domain("notify")
+async def test_async_get_domains(async_cached_client: AsyncClient) -> None:
+    notify = await async_cached_client.get_domain("notify")
     assert notify is not None
     assert notify.domain_id == "notify"
     with pytest.raises(AttributeError):
@@ -102,8 +100,8 @@ def test_entity_get_history(cached_client: Client) -> None:
         assert isinstance(state, State)
 
 
-async def test_async_entity_get_history(async_cached_client: Client) -> None:
-    entity = await async_cached_client.async_get_entity(group_id="sun", slug="sun")
+async def test_async_entity_get_history(async_cached_client: AsyncClient) -> None:
+    entity = await async_cached_client.get_entity(group_id="sun", slug="sun")
     assert entity is not None
     history = await entity.async_get_history()
     assert history is not None
@@ -132,8 +130,8 @@ def test_domain_from_json_with_client_missing_keys(cached_client: Client) -> Non
         Domain.from_json_with_client({"foo": "bar"}, cached_client)
 
 
-async def test_async_entity_get_history_none(async_cached_client: Client) -> None:
-    entity = await async_cached_client.async_get_entity(group_id="sun", slug="red_sun")
+async def test_async_entity_get_history_none(async_cached_client: AsyncClient) -> None:
+    entity = await async_cached_client.get_entity(group_id="sun", slug="red_sun")
     assert entity is not None
     history = await entity.async_get_history(
         start_timestamp=datetime(2015, 1, 1), end_timestamp=datetime(2020, 1, 1)
